+++ src/harlequin/exception.py
@@ -9,18 +9,21 @@ class HarlequinError(Exception):
 class HarlequinConnectionError(HarlequinError):
     def __init__(self, msg: str, title: str = "") -> None:
         super().__init__(msg)
+        self.msg = msg
         self.title = title
 
 
 class HarlequinCopyError(HarlequinError):
     def __init__(self, msg: str, title: str = "") -> None:
         super().__init__(msg)
+        self.msg = msg
         self.title = title
 
 
 class HarlequinQueryError(HarlequinError):
     def __init__(self, msg: str, title: str = "") -> None:
         super().__init__(msg)
+        self.msg = msg
         self.title = title